legal-content-system/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# PostgreSQL driver (optional, for production)
psycopg2-binary==2.9.9

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML
anthropic==0.18.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# WordPress Integration
requests==2.31.0

# Utilities
python-slugify==8.0.2
markdown==3.5.2

# Security
cryptography==42.0.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Date/Time
python-dateutil==2.8.2

# Logging and Monitoring
loguru==0.7.2

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality (dev)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
